create request_logs app, use in classifiers
