+++ temba/classifiers/types/wit/type.py
@@ -1,5 +1,6 @@
-from ...models import ClassifierType, Intent, ClassifierLog
+from ...models import ClassifierType, Intent
 from .views import ConnectView
+from temba.request_logs.models import HTTPLog
 
 import requests
 from django.utils import timezone
@@ -41,7 +42,9 @@ class WitType(ClassifierType):
         response = requests.get(cls.INTENT_URL, headers={"Authorization": f"Bearer {access_token}"})
         elapsed = (timezone.now() - start).total_seconds() * 1000
 
-        log = ClassifierLog.from_response(classifier, cls.INTENT_URL, response, "Synced Intents", "Syncing Error")
+        log = HTTPLog.from_response(
+            HTTPLog.INTENTS_SYNCED, cls.INTENT_URL, response, "Synced Intents", "Syncing Error", classifier=classifier
+        )
         log.request_time = elapsed
         logs.append(log)
