+++ temba/request_logs/urls.py
@@ -1,0 +1,5 @@
+from django.conf.urls import include, url
+
+from .views import HTTPLogCRUDL
+
+urlpatterns = [url(r"^", include(HTTPLogCRUDL().as_urlpatterns()))]
