+++ temba/settings_common.py
@@ -236,6 +236,7 @@ INSTALLED_APPS = (
     "temba.assets",
     "temba.auth_tweaks",
     "temba.api",
+    "temba.request_logs",
     "temba.classifiers",
     "temba.dashboard",
     "temba.public",
@@ -526,8 +527,6 @@ GROUP_PERMISSIONS = {
         "archives.archive.*",
         "campaigns.campaign.*",
         "campaigns.campaignevent.*",
-        "classifiers.classifierlog_list",
-        "classifiers.classifierlog_read",
         "classifiers.classifier_api",
         "classifiers.classifier_read",
         "classifiers.classifier_delete",
@@ -634,6 +633,8 @@ GROUP_PERMISSIONS = {
         "policies.policy_read",
         "policies.policy_list",
         "policies.policy_give_consent",
+        "request_logs.httplog_list",
+        "request_logs.httplog_read",
         "templates.template_api",
         "triggers.trigger.*",
     ),
