+++ templates/channels/channel_claim.haml
@@ -29,7 +29,7 @@
                 .btn-container
                   .claim-glyph{class:"{{ ch_type.icon }}"}
                   %p.claim-text= ch_type.name
-            .span9
+            .span8
               .claim-description
                 %div
                   {% include ch_type.get_claim_blurb %}
@@ -48,7 +48,7 @@
                 %p.claim-text
                   -trans "Android Phone"
 
-          .span9
+          .span8
             .claim-description
               %div
                 -blocktrans
@@ -69,7 +69,7 @@
                 %p.claim-text
                   -trans "Android Phone"
 
-          .span9
+          .span8
             .claim-description
               %div
                 -blocktrans
@@ -85,7 +85,7 @@
               .btn-container
                 .claim-glyph{class:"{{ ch_type.icon }}"}
                 %p.claim-text= ch_type.name
-          .span9
+          .span8
             .claim-description
               %div
                 {% include ch_type.get_claim_blurb %}
@@ -101,7 +101,7 @@
               .btn-container
                 .claim-glyph{class:"{{ ch_type.icon }}"}
                 %p.claim-text= ch_type.name
-          .span9
+          .span8
             .claim-description
               %div
                 {% include ch_type.get_claim_blurb %}
@@ -117,7 +117,7 @@
               .btn-container
                 .claim-glyph{class:"{{ ch_type.icon }}"}
                 %p.claim-text= ch_type.name
-          .span9
+          .span8
             .claim-description
               %div
                 {% include ch_type.get_claim_blurb %}
@@ -257,6 +257,6 @@
       text-decoration: none;
     }
 
-    .claim-row .span9 {
+    .claim-row .span8 {
       margin-left: 0px;
     }
