+++ templates/classifiers/classifier_connect.haml
@@ -27,7 +27,7 @@
               .btn-container
                 .classifier-glyph{class:"{{ cl_type.icon }}"}
                 %p.classifier-text= cl_type.name
-          .span9
+          .span8
             .classifier-description
               %div
                 {% include cl_type.get_connect_blurb %}
@@ -77,6 +77,6 @@
       margin-bottom: 10px;
     }
 
-    .classifier-row .span9 {
+    .classifier-row .span8 {
       margin-left: 0px;
     }
