+++ templates/classifiers/classifier_read.haml
@@ -17,7 +17,7 @@
 
 
   .header-margin
-    %a.btn.btn-tiny(href="{% url 'classifiers.classifierlog_list' classifier.uuid %}")
+    %a.btn.btn-tiny(href="{% url 'request_logs.httplog_list' 'classifier' classifier.uuid %}")
       -trans "Classifier Log"
 
 -block content
